Move pipe(2) based waker to its own file
Also here the change from pub to pub(crate) was made.

No functional changes.
